Tests for entrypoint moved into own file
